requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
ultralytics==8.3.170
opencv-python-headless==4.12.0.88
python-multipart==0.0.20
pillow==11.3.0
numpy==2.2.6
